🧹 Reorder creator file structure
