+++ creator/src/components/elements/element-image.js
@@ -3,7 +3,7 @@ import React from 'react';
 import produce from 'immer';
 import { makeStyles } from '@material-ui/core/styles';
 
-import Editable from '../shared/editable';
+import Editable from '../common/editable';
 
 const useStyles = makeStyles((theme) => ({
   image: {
