+++ creator/src/components/elements/element-label.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import produce from 'immer';
 
-import Editable from '../shared/editable';
+import Editable from '../common/editable';
 
 function ElementLabel({ structure, onUpdate }) {
